+++ ParaView/ParaView/vtkPVAxesActor.h
@@ -70,8 +70,8 @@ public:
   // Description:
   // Get the bounds for this Actor as (Xmin,Xmax,Ymin,Ymax,Zmin,Zmax). (The
   // method GetBounds(double bounds[6]) is available from the superclass.)
-  void GetBounds(float bounds[6]);
-  float *GetBounds();
+  void GetBounds(double bounds[6]);
+  double *GetBounds();
 
   // Description:
   // Get the actors mtime plus consider its properties and texture if set.
